requirements.txt
opencv-python
numpy
torch
torchvision
ultralytics
flask
pydantic
pyyaml
